backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=2.0
django-cors-headers>=3.15
python-dotenv>=1.0
Pillow>=9.0
whitenoise>=6.0
gunicorn>=21.0
